fix(moe): move the default dtype into tt_routed_expert, drop the one-constant module
Review: a whole module for a single constant is heavy. It now lives with the
class that owns the cache it describes, and utils/expert_dtypes.py is gone.

No import cycle: tt_routed_expert imports only lightweightmodule,
utility_functions and init_helpers, and tt_moe already imported TtRoutedExpert
from it. Behaviour is unchanged -- same value, same call sites.
